feat: add typescript and rename all files src directory
