+++ src/app.ts
@@ -2,9 +2,9 @@ const express = require('express');
 const swaggerUI = require('swagger-ui-express');
 const path = require('path');
 const YAML = require('yamljs');
-const userRouter = require('./resources/users/user.router');
-const boardRouter = require('./resources/boards/board.router');
-const taskRouter = require('./resources/tasks/task.router');
+const userRouter = require('./resources/users/user.router.ts');
+const boardRouter = require('./resources/boards/board.router.ts');
+const taskRouter = require('./resources/tasks/task.router.ts');
 
 const app = express();
