+++ src/resources/boards/board.router.ts
@@ -1,5 +1,5 @@
 const router = require('express').Router();
-const boardsService = require('./board.service');
+const boardsService = require('./board.service.ts');
 
 router.route('/').get(async (req, res) => {
   const boards = await boardsService.getAll();
