+++ src/resources/tasks/task.service.ts
@@ -1,4 +1,4 @@
-const tasksRepo = require('./task.memory.repository');
+const tasksRepo = require('./task.memory.repository.ts');
 
 /**
  * Task service module
