+++ src/resources/users/user.router.ts
@@ -1,6 +1,6 @@
 const router = require('express').Router();
-const User = require('./user.model');
-const usersService = require('./user.service');
+const User = require('./user.model.ts');
+const usersService = require('./user.service.ts');
 
 router.route('/').get(async (req, res) => {
   const users = await usersService.getAll();
