+++ src/resources/users/user.service.ts
@@ -1,4 +1,4 @@
-const usersRepo = require('./user.memory.repository');
+const usersRepo = require('./user.memory.repository.ts');
 
 /**
  * User service module
