+++ src/server.ts
@@ -1,5 +1,5 @@
-const { PORT } = require('./common/config');
-const app = require('./app');
+const { PORT } = require('./common/config.ts');
+const app = require('./app.ts');
 
 app.listen(PORT, () =>
   console.log(`App is running on http://localhost:${PORT}`)
